Correcciones para que compile correctamente.

+++ Nota.cpp
@@ -5,6 +5,6 @@ Nota::Nota(double valor){
     this->nota = valor;
 }
 double Nota::getNota(){return this->nota;}
-void Nota::setNota(int valor){this->nota = valor;}
+void Nota::setNota(double valor){this->nota = valor;}
 Nota::~Nota(){}
